Invoice extractor/requirement.txt
langchain
python-dotenv
chromadb
streamlit
PyPDF2
google-generativeai
